+++ plugins/engines/optix/cuda/renderer/Phong.cu
@@ -66,7 +66,7 @@ RT_PROGRAM void closest_hit_radiance()
     float3 world_geometric_normal = normalize( rtTransformNormal( RT_OBJECT_TO_WORLD, geometric_normal ) );
 
     float3 ffnormal = faceforward( world_shading_normal, -ray.direction, world_geometric_normal );
-    phongShade( Kd, Ka, Ks, Kr, Ko, refraction_index, phong_exp, ffnormal );
+    phongShade( Kd, Ka, Ks, Kr, Ko, refraction_index, phong_exp, ffnormal, ray.tmax );
 }
 
 
@@ -78,5 +78,5 @@ RT_PROGRAM void closest_hit_radiance_textured()
     float3 ffnormal = faceforward( world_shading_normal, -ray.direction, world_geometric_normal );
 
     const float3 Kd = make_float3( tex2D( diffuse_map, texcoord.x, texcoord.y ) );
-    phongShade( Kd, Ka, Ks, Kr, Ko, refraction_index, phong_exp, ffnormal );
+    phongShade( Kd, Ka, Ks, Kr, Ko, refraction_index, phong_exp, ffnormal, ray.tmax );
 }