+++ doc/ChangeLog.md
@@ -3,6 +3,9 @@ Changelog {#Changelog}
 
 # master
 
+* [#103](https://github.com/BlueBrain/Brayns/pull/103:
+  Added clipping planes (OSPRay engine only). --camera-type command line argument must
+  be set to activate the feature
 * [#95](https://github.com/BlueBrain/Brayns/pull/95):
   Upgraded to OSPRay 1.1.2
 * [#87](https://github.com/BlueBrain/Brayns/pull/87):